+++ CHANGES.md
@@ -1,5 +1,6 @@
 0.8.1
 -----
+* Celluloid::Pool lets you supervise a pool of actors
 * Celluloid#exclusive runs a high priority block which prevents other methods
   from executing